pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "keystone"
version = "0.1.0"
description = "DAG-based task orchestration daemon for ProjectKeystone"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.0",
    "nats-py>=2.7.0",
    "structlog>=24.1.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/keystone --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = [
    "src/keystone/__main__.py",
    "src/keystone/daemon.py",
]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["nats.*", "aiohttp.*"]
ignore_missing_imports = true

[tool.setuptools.packages.find]
where = ["src"]
